COMPLETE PROGRAM!
Now the game is completed! The AI works perfectly. It uses an non efficient algorith,
but capable of guessing the secrect combination in less than 12 movements.
Not always it achives that, so there is a possibility that the AI loses the game.
So: the AI is hard to win, but not invencible

Last step is: polish de code, doing this
- Declare which methods must be public or private
- Improve commentaries
- Explain the internal operation of the AI in the README file
- Pass the Rubocop through the code